+++ src/data.rs
@@ -5,8 +5,9 @@ use actix_http::error::{Error, ErrorInternalServerError};
 use actix_http::Extensions;
 use futures::{Async, Future, IntoFuture, Poll};
 
+use crate::dev::Payload;
 use crate::extract::FromRequest;
-use crate::service::ServiceFromRequest;
+use crate::request::HttpRequest;
 
 /// Application data factory
 pub(crate) trait DataFactory {
@@ -91,8 +92,8 @@ impl<T: 'static, P> FromRequest<P> for Data<T> {
     type Future = Result<Self, Error>;
 
     #[inline]
-    fn from_request(req: &mut ServiceFromRequest<P>) -> Self::Future {
-        if let Some(st) = req.request().config().extensions().get::<Data<T>>() {
+    fn from_request(req: &HttpRequest, _: &mut Payload<P>) -> Self::Future {
+        if let Some(st) = req.app_config().extensions().get::<Data<T>>() {
             Ok(st.clone())
         } else {
             Err(ErrorInternalServerError(
@@ -230,7 +231,7 @@ impl<T: 'static, P> FromRequest<P> for RouteData<T> {
     type Future = Result<Self, Error>;
 
     #[inline]
-    fn from_request(req: &mut ServiceFromRequest<P>) -> Self::Future {
+    fn from_request(req: &HttpRequest, _: &mut Payload<P>) -> Self::Future {
         if let Some(st) = req.route_data::<T>() {
             Ok(st.clone())
         } else {